+++ src/engine/graphics/vk/vulkan.h
@@ -21,8 +21,8 @@ public:
 };
 
 struct vkRenderableItem : public RenderableItem {
-  vkRenderableItem(Game::Vertex* vertices, size_t vcount, glm::uint16_t* indices, size_t icount, PIPELINE p);
-  void updateData(Game::Vertex* vertices, size_t vcount, glm::uint16_t* indices, size_t icount) override;
+  vkRenderableItem(Game::Vertex* vertices, uint32_t vcount, glm::uint16_t* indices, uint32_t icount, PIPELINE p);
+  void updateData(Game::Vertex* vertices, uint32_t vcount, glm::uint16_t* indices, uint32_t icount) override;
   ~vkRenderableItem() override;
   void setUniformModelMatrix(glm::mat4 m) override;
   void updateUniform() override;
